+++ core/viewmodel/src/main/kotlin/com/captures2024/soongan/core/viewmodel/profile/ProfileViewModel.kt
@@ -60,7 +60,7 @@ constructor(
         data object NavigateToNotification : Effect
 
         data class NavigateToHomePost(
-            val postId: Int,
+            val postId: Long,
         ) : Effect
 
         data object NavigateToRegistrationPost : Effect
@@ -78,7 +78,7 @@ constructor(
 
         data object LoadNextPage : Intent
 
-        data class OnClickPhoto(val postId: Int) : Intent
+        data class OnClickPhoto(val postId: Long) : Intent
 
         data object OnClickRegistrationText : Intent
